getting mongo working with async
