+++ src/main/java/the/alley/controller/ThymeleafController.java
@@ -1,9 +1,8 @@
 package the.alley.controller;
 
-import the.alley.db.UserRepo;
+import the.alley.model.UserRepo;
 import the.alley.utils.DarknessConstants;
 import the.alley.utils.Methods;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
