+++ src/main/java/the/alley/model/ItemsDB.java
@@ -1,4 +1,4 @@
-package the.alley.db;
+package the.alley.model;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,19 +6,18 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 import javax.persistence.*;
 
-//map tables
-
+//item table
 @Getter
 @Setter
 @Document
-@Table(name = "map")
-public class MapDB {
+@Table(name = "items")
+public class ItemsDB {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int id;
 	private String name;
 	private String description;
-	private Integer items;
-	private Integer npcs;
-	private Integer users;
+	private Integer attack;
+	private Integer defense;
+
 }
