+++ src/main/java/the/alley/model/ItemsRepo.java
@@ -1,4 +1,4 @@
-package the.alley.db;
+package the.alley.model;
 
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
