+++ src/main/java/the/alley/model/UserDB.java
@@ -1,23 +1,28 @@
-package the.alley.db;
+package the.alley.model;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import javax.persistence.*;
+//user objects
 
-//item table
 @Getter
 @Setter
 @Document
-@Table(name = "items")
-public class ItemsDB {
+@Table(name = "users")
+public class UserDB {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private int id;// test
 	private String name;
-	private String description;
+	private Integer lvl;
+	private Integer money;
+	private Integer exp;
 	private Integer attack;
 	private Integer defense;
+	private String description;
+	private Integer location;
+	private Integer hp;
 
 }
