+++ src/main/resources/application.properties
@@ -49,39 +49,34 @@ spring.jpa.show-sql=true
 #spring.datasource.driverClassName=org.mariadb.jdbc.Driver
 
 #mongo set up
-spring.data.mongodb.host=localhost  #Mongo server host.
-spring.data.mongodb.port=1919  #Mongo server port.
-spring.data.mongodb.repositories.enabled=true  #Enable Mongo repositories.
-spring.data.mongodb.uri=mongodb://localhost:1919/the_alley  #Mongo database URI. When set, host and port are ignored.
-spring.data.mongod.username=root
-spring.data.mongodb.password=password
-
-ffgfdgfgdvxz
-# Naming strategy
-# spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
-# spring.jpa.properties.hibernate.id.new_generator_mappings=true
-# Use spring.jpa.properties.* for Hibernate native properties (the prefix is
-# stripped before adding them to the entity manager)
-# The SQL dialect makes Hibernate generate better SQL for the chosen database
-# spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MariaDB103Dialect
-# Hibernate ddl auto (create, create-drop, update, validate)  spring.jpa.hibernate.ddl-auto=create-drop or validate
-# spring.jpa.hibernate.ddl-auto=create-drop
-# spring.jpa.show-sql=true
-# Naming strategy
-# spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
-# spring.jpa.properties.hibernate.id.new_generator_mappings=true
-# Use spring.jpa.properties.* for Hibernate native properties (the prefix is
-# stripped before adding them to the entity manager)
-# The SQL dialect makes Hibernate generate better SQL for the chosen database
-# spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MariaDB103Dialect
-
-#mongo db set up
 #spring.data.mongodb.authentication-database= # Authentication database name.
-#spring.data.mongodb.database=the_alley # Database name.
+spring.data.mongodb.database=the_alley
 #spring.data.mongodb.field-naming-strategy= # Fully qualified name of the FieldNamingStrategy to use.
-#spring.data.mongodb.grid-fs-database= # GridFS database name.
+#GridFS database name.
+#spring.data.mongodb.grid-fs-database=testing
 #spring.data.mongodb.password= sysadmin  # Login password of the mongo server.
 #spring.data.mongodb.username= welcome1 # Login user of the mongo server.
+spring.data.mongodb.host=shibby
+spring.data.mongodb.port=1919
+spring.data.mongodb.repositories.enabled=true
+#spring.data.mongodb.uri=mongodb://shibby:1919/testing
+spring.data.mongod.username=root
+spring.data.mongodb.password=password
+spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
+spring.jpa.properties.hibernate.id.new_generator_mappings=true
+# Use spring.jpa.properties.* for Hibernate native properties (the prefix is
+# stripped before adding them to the entity manager)
+# The SQL dialect makes Hibernate generate better SQL for the chosen database
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.AbstractHANADialect
+#(create, create-drop, update, validate)  spring.jpa.hibernate.ddl-auto=create-drop or validate
+
+# Naming strategy
+# spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
+# spring.jpa.properties.hibernate.id.new_generator_mappings=true
+# Use spring.jpa.properties.* for Hibernate native properties (the prefix is
+# stripped before adding them to the entity manager)
+# The SQL dialect makes Hibernate generate better SQL for the chosen database
+# spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MariaDB103Dialect
 
 #security
 #security.user.name=maestro
